+++ examples/maya_plugin/test/testImportMayaModel.py
@@ -65,6 +65,8 @@ class ImportMayaModelTestCase(unittest.TestCase):
     def setUpClass(cls):
         if not cls.plugins_loaded:
             standalone.initialize("bifrost-usd-import-maya-model")
+            cmds.loadPlugin("mayaUsdPlugin", quiet=True)
+            cmds.loadPlugin("bifrostGraph", quiet=True)
             cmds.loadPlugin("bifrostUSDExamples.py", quiet=True)
             cls.plugins_loaded = True